Combined all JS into server file
