+++ public/reserve.html
@@ -70,7 +70,7 @@
             $.post("/data", newreserve)
                 .then(function (data) {
                     console.log("New Reservation", data);
-                    alert("Atempting to add reservation...");
+                    // alert("Atempting to add reservation...");
                 });
         });
